requirements.txt
flask
groq
python-dotenv
gunicorn
psycopg2-binary
